requirements.txt
Flask~=3.1.1
python-docx==0.8.11       # Работа с DOCX файлами (основная функциональность)
docxtpl
PyYAML~=6.0.2
requests~=2.32.3